+++ src/main/java/product/option/Option.java
@@ -6,17 +6,17 @@ import java.math.BigDecimal;
  * Option Entity
  */
 public class Option {
-    private final Integer id;
+    private final Long id;
     private final String name;
     private final BigDecimal cost;
 
-    public Option(Integer id, String name, BigDecimal cost) {
+    public Option(Long id, String name, BigDecimal cost) {
         this.id = id;
         this.name = name;
         this.cost = cost;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     public String getName() {
